backend/requirements.txt
fastapi[standard]
uvicorn
click
sqlalchemy
psycopg2-binary
pydantic-settings
alembic
uuid6
python-dotenv
passlib[bcrypt]
# passlib 1.7.4 is not compatible with bcrypt 5.x
bcrypt<5
python-jose[cryptography]
python-multipart
email-validator

yfinance
